+++ cognite/client/_api/simulators/simulator_integrations.py
@@ -4,9 +4,9 @@ from typing import TYPE_CHECKING, Any
 
 from cognite.client._api_client import APIClient
 from cognite.client._constants import DEFAULT_LIMIT_READ
+from cognite.client.data_classes.simulators.filters import SimulatorIntegrationFilter
 from cognite.client.data_classes.simulators.simulators import (
     SimulatorIntegration,
-    SimulatorIntegrationFilter,
     SimulatorIntegrationList,
 )
 from cognite.client.utils._experimental import FeaturePreviewWarning
